docs/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic>=2.7.4
python-dotenv==1.0.0
jinja2==3.1.3
pandas==2.1.4
playwright==1.41.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dateutil==2.8.2
weasyprint==61.0
# Chart.js will be loaded via CDN in templates

# LangChain for multi-LLM support (pinned to compatible versions)
langchain>=0.1.20,<0.2.0
langchain-anthropic>=0.1.5,<0.2.0
langchain-openai>=0.1.7,<0.2.0
langchain-core>=0.1.53,<0.2.0
anthropic>=0.18.0

# Word document generation (optional - for .docx export)
# Install one of these options:
# Option 1 (recommended): python-docx + htmldocx
python-docx>=1.1.0
htmldocx>=0.0.5
# Option 2 (alternative): mammoth
# mammoth>=1.6.0

# Authentication & Database
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9  # PostgreSQL driver for Supabase
email-validator>=2.0.0  # Required for Pydantic EmailStr validation